bill day offset should be an integer instead of decimal

+++ db/payment_wentian_2.0-delta-schema.sql
@@ -25,5 +25,5 @@ CREATE TABLE `eh_payment_formula` (
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COMMENT='收费标准公式表';
 
 ALTER TABLE `eh_payment_charging_standards` ADD COLUMN `suggest_unit_price` DECIMAL(10,2) DEFAULT NULL COMMENT '建议单价';
-ALTER TABLE `eh_payment_charging_standards` ADD COLUMN `bill_item_month_offset` DECIMAL(10,2) DEFAULT NULL COMMENT '收费项产生时间偏离当前月的月数';
-ALTER TABLE `eh_payment_charging_standards` ADD COLUMN `bill_item_day_offset` DECIMAL(10,2) DEFAULT NULL COMMENT '收费项产生时间偏离当前月的日数';
+ALTER TABLE `eh_payment_charging_standards` ADD COLUMN `bill_item_month_offset` INTEGER DEFAULT NULL COMMENT '收费项产生时间偏离当前月的月数';
+ALTER TABLE `eh_payment_charging_standards` ADD COLUMN `bill_item_day_offset` INTEGER DEFAULT NULL COMMENT '收费项产生时间偏离当前月的日数';
